ethereum-contracts/requirements.txt
py-solc-x==0.5.0